BOOK-27 | Rewrite tests after merge
