+++ BookHub/BookHub/settings/base.py
@@ -138,3 +138,5 @@ REST_FRAMEWORK = {
 }
 
 COMMENT_SNIPPET_LENGTH = 30
+
+BOOK_LAST_COMMENTS = 3
